architect-serverless-python/requirements.txt
architect-functions==0.2.0
boto3==1.7.55